+++ R/plot.pwsd.R
@@ -7,6 +7,7 @@
 #' @param x an of object of class 'pwsd' or 'acf'
 #' @param main an overall title for the plot, if no string is supplied a default
 #'  title  will be populated. \emph{See} \code{\link[graphics]{title}}
+#' @param ylim a numeric of length 2 giving the y-axis limits for the plot
 #'
 #' @inheritParams pwsd
 #' @inheritDotParams base::plot
@@ -26,7 +27,7 @@
 #' b <- pwsd(sim, round = TRUE, correlogram = FALSE)
 #' plot(b)
 #'
-plot.pwsd <- function(x, c = NULL, main = NULL, ...) {
+plot.pwsd <- function(x, c = NULL, main = NULL, ylim = NULL, ...) {
 
   # Check class
   stopifnot(any(class(x) == "pwsd"))
@@ -38,25 +39,35 @@ plot.pwsd <- function(x, c = NULL, main = NULL, ...) {
   # Set c to input if not NULL
   if (is.null(c)) {
     c <- x$parameters[, "c"]
-  } else {c <- c}
+  }
+
+  # Set significance bands
+  rho_k_critical <- c * sqrt(log10(x$parameters[, "n"]) / x$parameters[, "n"])
 
   # Plot acf
   j <- 1
   for (i in acf) {
 
+    # Set plot title
     if (is.null(main)) {
       main = paste0("Correlogram for: ", names[j])
-    } else {
-      main = main
     }
 
-    # Plot ACT
+    # Set y-axis limits
+    ylim <- range(i$acf, rho_k_critical, -rho_k_critical)
+
+    # Plot ACF
     plot(i,
-      ci = stats::pnorm(c),
-      ci.col =  "darkmagenta",
+      ci = 0,
       xlab = "Lag (k)",
       main = main,
+      ylim = ylim,
       ...)
+
+    # Add implied significance bands
+    graphics::abline(h = c(rho_k_critical, -rho_k_critical),
+      col = "darkmagenta",
+      lty = "dotdash")
 
     # Name Counter
     j <- j + 1