+++ NEWS.md
@@ -1,5 +1,7 @@
 # blocklength 0.1.3.9004
 
+* A new vignette has been included on tuning and diagnosing problematic output from the selection functions
+
 * the argument `sub_block_length = ` of `hhj()` has been changed to `sub_sample = ` to avoid confusion with the other tuning parameter, `pilot_block_length()`
 
 * `rho.k.critical` has been changed to snake case `rho_k_critical` in the `$parameters` matrix from output of class 'pwsd' objects from `pwsd()`
@@ -8,6 +10,7 @@
 
 * `plot.pwsd()` now includes "darkmagenta" significance lines to match `pwsd()`
 
+* `plot.pwsd()` now explicity includes an option to customize title with `main = `
 
 # blocklength 0.1.3