+++ NEWS.md
@@ -1,8 +1,18 @@
 # blocklength 0.1.3.9004
 
-* A new vignette has been included on tuning and diagnosing problematic output from the selection functions
+
+## Bug Fixes
+
+* The estimation of `m_hat` within `pwsd()` was indexing the first insignificant lag of the correlation structure rather than the first *significant* lag prior to the consecutive run of insignificant lags.
+
+* The correlogram output of `pwsd()` and `plot.pwsd()` now output significance bands that match `rho_k_critical` exactly. Prior to this, the significant bands on the correlogram output were generated using the `ci = ` arguement which led to misleading graphical representations of the implied hypothesis test of the PWSD method.
 
 * the argument `sub_block_length = ` of `hhj()` has been changed to `sub_sample = ` to avoid confusion with the other tuning parameter, `pilot_block_length()`
+
+
+## Minor Changes
+
+* A new vignette has been included on tuning and diagnosing problematic output from the selection functions!
 
 * `rho.k.critical` has been changed to snake case `rho_k_critical` in the `$parameters` matrix from output of class 'pwsd' objects from `pwsd()`
 
@@ -10,7 +20,8 @@
 
 * `plot.pwsd()` now includes "darkmagenta" significance lines to match `pwsd()`
 
-* `plot.pwsd()` now explicity includes an option to customize title with `main = `
+* `plot.pwsd()` now explicitly includes an option to customize title with `main = `
+
 
 # blocklength 0.1.3