+++ NEWS.md
@@ -14,6 +14,8 @@
 
 * A new vignette has been included on tuning and diagnosing problematic output from the selection functions!
 
+* `pwsd()` now includes a new arguement to overide the implied hypothesis test by setting `m_hat` directly
+
 * `rho.k.critical` has been changed to snake case `rho_k_critical` in the `$parameters` matrix from output of class 'pwsd' objects from `pwsd()`
 
 * if `subsample = ` of `hhj()` is set directly, it is now rounded to the nearest whole number.