+++ NEWS.md
@@ -1,4 +1,4 @@
-# blocklength 0.1.3.9004
+# blocklength 0.1.3.9005
 
 
 ## Bug Fixes